[GH-31] Revert configurable php.ini changes
The php cookbook adds a cli php.ini that we can change. There are
php.ini files for apache2 and cgi that we don't have any control
over. The workaround of adding a cgi.php.ini file to the ext_conf_dir
doesn't actually work because the ext_conf_dir is scanned by
apache2, cgi, and cli. We don't want cgi.php.ini to be included
with the cli environment. Need to try another approach.