[GH-65] Run squid recipe after xforty
A new squid package was released. Running the squid recipe before
xforty meant that we didn't have an up-to-date apt cache. The
apt-get update command is run in the xforty recipe. As a result,
the squid recipe was trying to install an old version of the
package which doesn't exist in the mirrors anymore. Making sure
the xforty recipe is run first resolves the issue.